+++ CHANGES.rst
@@ -1,3 +1,5 @@
+* Added SubWidgets to allow you to iterate over invdidiual elements which make
+  up a widget -- currently only used by RadioSelect.
 * Changed MultiValueField to run any validators it was given.
 * Changed URL and email address validators to handle IDNA domains.
 * Changed CheckboxInput to correctly handle ``0`` as a value.